chore: split score params/thresholds files
