+++ test/peerScoreParams.spec.js
@@ -1,57 +1,8 @@
 const { expect } = require('chai')
 const {
-  createPeerScoreThresholds, validatePeerScoreThresholds,
   createTopicScoreParams, validateTopicScoreParams,
   createPeerScoreParams, validatePeerScoreParams
-} = require('../src/score/scoreParams')
-
-describe('PeerScoreThresholds validation', () => {
-  it('should throw on invalid PeerScoreThresholds', () => {
-    expect(() => validatePeerScoreThresholds(
-      createPeerScoreThresholds({
-        gossipThreshold: 1
-      })
-    )).to.throw
-    expect(() => validatePeerScoreThresholds(
-      createPeerScoreThresholds({
-        publishThreshold: 1
-      })
-    )).to.throw
-    expect(() => validatePeerScoreThresholds(
-      createPeerScoreThresholds({
-        gossipThreshold: -1,
-        publishThreshold: 0
-      })
-    )).to.throw
-    expect(() => validatePeerScoreThresholds(
-      createPeerScoreThresholds({
-        gossipThreshold: -1,
-        publishThreshold: -2 
-      })
-    )).to.throw
-    expect(() => validatePeerScoreThresholds(
-      createPeerScoreThresholds({
-        acceptPXThreshold: -1
-      })
-    )).to.throw
-    expect(() => validatePeerScoreThresholds(
-      createPeerScoreThresholds({
-        opportunisticGraftThreshold: -1
-      })
-    )).to.throw
-  })
-  it('should not throw on valid PeerScoreThresholds', () => {
-    expect(() => validatePeerScoreThresholds(
-      createPeerScoreThresholds({
-        gossipThreshold: -1,
-        publishThreshold: -2,
-        graylistThreshold: -3,
-        acceptPXThreshold: 1,
-        opportunisticGraftThreshold: 2
-      })
-    )).to.not.throw
-  })
-})
+} = require('../src/score')
 
 describe('TopicScoreParams validation', () => {
   it('should throw on invalid TopicScoreParams', () => {
